requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-WTF==1.0.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
sib-api-v3-sdk==7.3.0
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
